[FOLD] Consolidate files
