+++ src/beast/beast/module/core/native/linux_SystemStats.cpp
@@ -30,6 +30,25 @@ void outputDebugString (std::string const& text)
 }
 
 //==============================================================================
+bool beast_isRunningUnderDebugger()
+{
+    static char testResult = 0;
+
+    if (testResult == 0)
+    {
+        testResult = (char) ptrace (PT_TRACE_ME, 0, 0, 0);
+
+        if (testResult >= 0)
+        {
+            ptrace (PT_DETACH, 0, (caddr_t) 1, 0);
+            testResult = 1;
+        }
+    }
+
+    return testResult < 0;
+}
+
+//==============================================================================
 std::string getComputerName()
 {
     char name [256] = { 0 };
